Fix bug where packet with ID 0 was rejected
First packet ever received by nufw was never authenticated, packets
with id 0 was rejected.

+++ src/nufw/nufw.h
@@ -177,7 +177,7 @@ int auth_request_send(uint8_t type, struct queued_pckt *pckt);
 void close_tls_session();
 void shutdown_tls();
 
-unsigned long padd(packet_idl * packet);
+int padd(packet_idl * packet);
 int psearch_and_destroy(uint32_t packet_id, uint32_t * mark);
 void clear_packet_list();
 void clean_old_packets();
